:truck: refactor: directorize guide message and category (#21)
