+++ src/commands/help.ts
@@ -1,4 +1,4 @@
-import {embedWelcome} from '../bot/guideMessage.js';
+import {embedWelcome} from '../embedMessage/guideMessage.js';
 import {Message} from "discord.js";
 
 export async function execute(message: Message){
